update test implementation
